refactor: Cart Product Detail 코드 분리 (#231)
