--- itda-front/src/components/ProductDetail/mocks.ts
@@ -1,134 +1,0 @@
-import { IReview } from "types/ProductDetailTypes";
-
-export const reviewMock: IReview[] = [
-  {
-    id: 1,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요!",
-  },
-  {
-    id: 2,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요!78463234",
-  },
-  {
-    id: 3,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요!444444444444444444",
-  },
-  {
-    id: 4,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요!765765",
-  },
-  {
-    id: 5,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요!7897",
-  },
-  {
-    id: 6,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요1212121212!",
-  },
-  {
-    id: 7,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요!23423423",
-  },
-  {
-    id: 8,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요!",
-  },
-  {
-    id: 9,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요22!",
-  },
-  {
-    id: 10,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요333!",
-  },
-  {
-    id: 11,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: ["https://ifh.cc/g/gFFnTG.jpg", "https://ifh.cc/g/gFFnTG.jpg"],
-    contents: "신선하고 맛있어요4444!",
-  },
-  {
-    id: 12,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요!",
-  },
-  {
-    id: 13,
-    writer: {
-      name: "cro**",
-      imgUrl: "https://ifh.cc/g/gFFnTG.jpg",
-    },
-    date: "2021-07-22",
-    image: [],
-    contents: "신선하고 맛있어요5555!",
-  },
-];
